generalize a little the explore part
Generalize the explore part adding a single function for printing and a single function for plotting the word/char distributions (before I have one function for printing and one for plotting for both word count or char count).
modified also the explore function -> now the function recieve only the three dataframes already modified in the main. This script probably could be ok also for other kinds of data, the only thing to matter is to change line 243 and line 244 ('tweet' with the name of the column where there are stored the text data)

Until now the read function is useful only if the data are already split in three different csv files. I want to modify this function in order to be good also if the data are stored in only one single csv file.