Divided e fixed read function in multiple functions
Divided read function in mulitple functions: now it's possible to read different kinds of dataset (the datasets don't have to be already split, but now a split function eventually do the job). Fixed che multiple occurencies problem while reading with a function (handle_multiple_occurencies): there was a problem with file with name such as: "constrain_valid.csv", this kind of file contain the "train" word but of course it's the validation dataset. The problem was that previous code read that dataset as train. Now this problem should be fixed.
In the config file now there are also train, valid, test fraction to split the datataset if the dataset is not already split (so it's not divided in three different files).

Finally, added a function to clean the raws dataframes: in this way it will be much easier to automatize the analysis for different datasets.
In the config file there is "column_name_text" which is the name of the column where there are the text data, and the column_name_label which is the name of the column where there are the labels.