Little changes
Little changes: in vectorize_data.py now the vocabulary takes a list / tuple of sentences as input. In this way it is possible to have a complete vocabulary (also considering the test dataset)
In preprocess.py: added an initial if statement to test if the data are already split -> if yes: no split; if not: split with the fractions given by config file.
In the other files very little meaningful changes.