Modify the folder structure and include the logistic regressor
Modify the folder structure: now it is possible to do different analysis on different datasets. The data must be stored in a subfolder inside the 'datasets' folder and the name of the subfolder corresponds to the analysis the user wants to do (for example: covid, means that the user wants to analyze data regarding covid, and the subfolder name inside the folder "datasets" is precisely "covid") The analysis name must be set in the configuration file.
For the processed data the same changes have been done. Now the processed data are saved inside the subfolder contained in "preprocessed_datasets". The name of the subfolder is given by the analysis value from the configuration file. Until now it's possible to do analysis only on covid dataset, but i will implement also other datasets).

Finally I moved the train of the logistic regressor inside the train file, so mlexecution.py could be removed.