Added the evaluation script
Added the evaluation script and the plotting module for the confusion matrix and the history of NN.
Need to split the different evaluations for the different models in two functions.
Need also to save the word2vec model, in order to be much faster when evaluate the logistic regressor model.

FIXED: logistic regressor model had not a determinist behaviour, because of the behaviour of the word2vec model. In order to have a deterministic behaviour I changed num of workers (now set to 1) and set the seed not set. Now it should be OK