Added plots for the results
Added the possibility to save the plots (confusion matrix and history of neural network) in a folder called "plots" inside a subfolder named as the analysis the user is doing.

Now the word2vec model is saved after the training, so it's faster to load the model and do the evaluation part.

In evaluate.py : split the evaluation of the two models in two different functions (evaluate_logistic_regression and evaluate_neural_network).