Added some test functions for vectorize part
Added test functions for;
1) init_vector_layer
2) MeanEmbeddingVectorizer
3) vectorize_X_data_lr
4) flatten_unique_voc (new function for flattening and getting the unique words from the vocabulary)

Fixed the get_vocabulary function: now the function takes as input a list of string same as before, but the output now is a list of lists of sentence tokenized. In this way this vocabulary is good to fit for the Word2Vec model. 
For the tensorflow vectorization layer, now there is a new function called "flatten_unique_voc", which takes as input the previous vocabulary (list of lists of string). The output of this function is a flatten list with all the unique words. The output is good to fit as input for the init_vector_layer function.