Finished first tutorial
Finished the first tutorial: dataset selected is spam and in the tutorial there is the train and evaluation of the logistic regressor.

Removed the find_init_columns from the library, this is a function inside the scripts, so if a user wants to add one dataset it is just necessary to modify this function from script (so the user has not the need to modify the library).